+++ app/src/main/cpp/jni/animationTrack.inl
@@ -18,7 +18,7 @@
 #include "javax_microedition_m3g_AnimationTrack.h"
 
 JNIEXPORT jint JNICALL Java_javax_microedition_m3g_AnimationTrack__1getTargetProperty
-(JNIEnv* aEnv, jclass, jint aHandle)
+(JNIEnv* aEnv, jclass, jlong aHandle)
 {
     M3G_DO_LOCK
     jint props = (jint)m3gGetTargetProperty((M3GAnimationTrack)aHandle);
@@ -26,35 +26,35 @@ JNIEXPORT jint JNICALL Java_javax_microedition_m3g_AnimationTrack__1getTargetPro
     return props;
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_AnimationTrack__1getSequence
-(JNIEnv* aEnv, jclass, jint aHandle)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_AnimationTrack__1getSequence
+(JNIEnv* aEnv, jclass, jlong aHandle)
 {
     M3G_DO_LOCK
-    jint sequence = (jint)m3gGetSequence((M3GAnimationTrack)aHandle);
+    jlong sequence = (jlong)m3gGetSequence((M3GAnimationTrack)aHandle);
     M3G_DO_UNLOCK(aEnv)
     return sequence;
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_AnimationTrack__1ctor
-(JNIEnv* aEnv, jclass, jint aM3g, jint aHSequence, jint aProperty)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_AnimationTrack__1ctor
+(JNIEnv* aEnv, jclass, jlong aM3g, jlong aHSequence, jint aProperty)
 {
     M3G_DO_LOCK
-    jint handle = (jint)m3gCreateAnimationTrack((M3GInterface)aM3g, (M3GKeyframeSequence)aHSequence, (M3Gint)aProperty);
+    jlong handle = (jlong)m3gCreateAnimationTrack((M3GInterface)aM3g, (M3GKeyframeSequence)aHSequence, (M3Gint)aProperty);
     M3G_DO_UNLOCK(aEnv)
     return handle;
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_AnimationTrack__1getController
-(JNIEnv* aEnv, jclass, jint aHandle)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_AnimationTrack__1getController
+(JNIEnv* aEnv, jclass, jlong aHandle)
 {
     M3G_DO_LOCK
-    jint controller = (jint)m3gGetController((M3GAnimationTrack)aHandle);
+    jlong controller = (jlong)m3gGetController((M3GAnimationTrack)aHandle);
     M3G_DO_UNLOCK(aEnv)
     return controller;
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_AnimationTrack__1setController
-(JNIEnv* aEnv, jclass, jint aHandle, jint aHController)
+(JNIEnv* aEnv, jclass, jlong aHandle, jlong aHController)
 {
     M3G_DO_LOCK
     m3gSetController((M3GAnimationTrack)aHandle, (M3GAnimationController)aHController);
