+++ app/src/main/cpp/jni/appearance.inl
@@ -18,33 +18,33 @@
 #include "javax_microedition_m3g_Appearance.h"
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setLayer
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aLayer)
+(JNIEnv* aEnv,   jclass, jlong aHApp, jint aLayer)
 {
     M3G_DO_LOCK
     m3gSetLayer((M3GAppearance)aHApp, (M3Gint)aLayer);
     M3G_DO_UNLOCK(aEnv)
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getCompositingMode
-(JNIEnv* aEnv,   jclass, jint aHApp)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_Appearance__1getCompositingMode
+(JNIEnv* aEnv,   jclass, jlong aHApp)
 {
     M3G_DO_LOCK
-    jint compMode = (jint)m3gGetCompositingMode((M3GAppearance)aHApp);
+    jlong compMode = (jlong)m3gGetCompositingMode((M3GAppearance)aHApp);
     M3G_DO_UNLOCK(aEnv)
     return compMode;
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getFog
-(JNIEnv* aEnv,   jclass, jint aHApp)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_Appearance__1getFog
+(JNIEnv* aEnv,   jclass, jlong aHApp)
 {
     M3G_DO_LOCK
-    jint fog = (jint)m3gGetFog((M3GAppearance)aHApp);
+    jlong fog = (jlong)m3gGetFog((M3GAppearance)aHApp);
     M3G_DO_UNLOCK(aEnv)
     return fog;
 }
 
 JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getLayer
-(JNIEnv* aEnv,   jclass, jint aHApp)
+(JNIEnv* aEnv,   jclass, jlong aHApp)
 {
     M3G_DO_LOCK
     jint layer = (jint)m3gGetLayer((M3GAppearance)aHApp);
@@ -53,7 +53,7 @@ JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getLayer
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setCompositingMode
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aHMode)
+(JNIEnv* aEnv,   jclass, jlong aHApp, jlong aHMode)
 {
     M3G_DO_LOCK
     m3gSetCompositingMode((M3GAppearance)aHApp, (M3GCompositingMode)aHMode);
@@ -61,50 +61,50 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setCompositingMo
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setPolygonMode
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aHMode)
+(JNIEnv* aEnv,   jclass, jlong aHApp, jlong aHMode)
 {
     M3G_DO_LOCK
     m3gSetPolygonMode((M3GAppearance)aHApp, (M3GPolygonMode)aHMode);
     M3G_DO_UNLOCK(aEnv)
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getPolygonMode
-(JNIEnv* aEnv,   jclass, jint aHApp)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_Appearance__1getPolygonMode
+(JNIEnv* aEnv,   jclass, jlong aHApp)
 {
     M3G_DO_LOCK
-    jint polyMode = (jint)m3gGetPolygonMode((M3GAppearance)aHApp);
+    jlong polyMode = (jlong)m3gGetPolygonMode((M3GAppearance)aHApp);
     M3G_DO_UNLOCK(aEnv)
     return polyMode;
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setTexture
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aUnit, jint aHTex)
+(JNIEnv* aEnv,   jclass, jlong aHApp, jint aUnit, jlong aHTex)
 {
     M3G_DO_LOCK
     m3gSetTexture((M3GAppearance)aHApp, (M3Gint)aUnit, (M3GTexture)aHTex);
     M3G_DO_UNLOCK(aEnv)
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1ctor
-(JNIEnv* aEnv,   jclass, jint aM3g)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_Appearance__1ctor
+(JNIEnv* aEnv,   jclass, jlong aM3g)
 {
     M3G_DO_LOCK
-    jint handle = (jint)m3gCreateAppearance((M3GInterface)aM3g);
+    jlong handle = (jlong)m3gCreateAppearance((M3GInterface)aM3g);
     M3G_DO_UNLOCK(aEnv)
     return handle;
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getTexture
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aUnit)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_Appearance__1getTexture
+(JNIEnv* aEnv,   jclass, jlong aHApp, jint aUnit)
 {
     M3G_DO_LOCK
-    jint texture = (jint)m3gGetTexture((M3GAppearance)aHApp, (M3Gint)aUnit);
+    jlong texture = (jlong)m3gGetTexture((M3GAppearance)aHApp, (M3Gint)aUnit);
     M3G_DO_UNLOCK(aEnv)
     return texture;
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setFog
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aHFog)
+(JNIEnv* aEnv,   jclass, jlong aHApp, jlong aHFog)
 {
     M3G_DO_LOCK
     m3gSetFog((M3GAppearance)aHApp, (M3GFog)aHFog);
@@ -112,18 +112,18 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setFog
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_Appearance__1setMaterial
-(JNIEnv* aEnv,   jclass, jint aHApp, jint aHMaterial)
+(JNIEnv* aEnv,   jclass, jlong aHApp, jlong aHMaterial)
 {
     M3G_DO_LOCK
     m3gSetMaterial((M3GAppearance)aHApp, (M3GMaterial)aHMaterial);
     M3G_DO_UNLOCK(aEnv)
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_Appearance__1getMaterial
-(JNIEnv* aEnv,   jclass, jint aHApp)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_Appearance__1getMaterial
+(JNIEnv* aEnv,   jclass, jlong aHApp)
 {
     M3G_DO_LOCK
-    jint material = (jint)m3gGetMaterial((M3GAppearance)aHApp);
+    jlong material = (jlong)m3gGetMaterial((M3GAppearance)aHApp);
     M3G_DO_UNLOCK(aEnv)
     return material;
 }
