+++ app/src/main/cpp/jni/compositingMode.inl
@@ -18,7 +18,7 @@
 #include "javax_microedition_m3g_CompositingMode.h"
 
 JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isDepthTestEnabled
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jboolean enabled = (jboolean)m3gIsDepthTestEnabled((M3GCompositingMode)aHCompositingMode);
@@ -27,7 +27,7 @@ JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isDepth
 }
 
 JNIEXPORT jint JNICALL Java_javax_microedition_m3g_CompositingMode__1getBlending
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jint blending = (jint)m3gGetBlending((M3GCompositingMode)aHCompositingMode);
@@ -36,7 +36,7 @@ JNIEXPORT jint JNICALL Java_javax_microedition_m3g_CompositingMode__1getBlending
 }
 
 JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isAlphaWriteEnabled
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jboolean enabled = (jboolean)m3gIsAlphaWriteEnabled((M3GCompositingMode)aHCompositingMode);
@@ -45,7 +45,7 @@ JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isAlpha
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1enableDepthWrite
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jboolean aEnable)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jboolean aEnable)
 {
     M3G_DO_LOCK
     m3gEnableDepthWrite((M3GCompositingMode)aHCompositingMode, (M3Gbool)aEnable);
@@ -53,7 +53,7 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1enableDepth
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setAlphaThreshold
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jfloat aThreshold)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jfloat aThreshold)
 {
     M3G_DO_LOCK
     m3gSetAlphaThreshold((M3GCompositingMode)aHCompositingMode, (M3Gfloat)aThreshold);
@@ -61,7 +61,7 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setAlphaThr
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1enableColorWrite
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jboolean aEnable)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jboolean aEnable)
 {
     M3G_DO_LOCK
     m3gEnableColorWrite((M3GCompositingMode)aHCompositingMode, (M3Gbool)aEnable);
@@ -69,7 +69,7 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1enableColor
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setDepthOffset
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jfloat aFactor, jfloat aUnits)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jfloat aFactor, jfloat aUnits)
 {
     M3G_DO_LOCK
     m3gSetDepthOffset((M3GCompositingMode)aHCompositingMode, (M3Gfloat)aFactor, (M3Gfloat)aUnits);
@@ -77,24 +77,24 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setDepthOff
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1enableDepthTest
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jboolean aEnable)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jboolean aEnable)
 {
     M3G_DO_LOCK
     m3gEnableDepthTest((M3GCompositingMode)aHCompositingMode, (M3Gbool)aEnable);
     M3G_DO_UNLOCK(aEnv)
 }
 
-JNIEXPORT jint JNICALL Java_javax_microedition_m3g_CompositingMode__1ctor
-(JNIEnv* aEnv, jclass, jint aM3g)
+JNIEXPORT jlong JNICALL Java_javax_microedition_m3g_CompositingMode__1ctor
+(JNIEnv* aEnv, jclass, jlong aM3g)
 {
     M3G_DO_LOCK
-    jint handle = (jint)m3gCreateCompositingMode((M3GInterface)aM3g);
+    jlong handle = (jlong)m3gCreateCompositingMode((M3GInterface)aM3g);
     M3G_DO_UNLOCK(aEnv)
     return handle;
 }
 
 JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isColorWriteEnabled
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jboolean enabled = (jboolean)m3gIsColorWriteEnabled((M3GCompositingMode)aHCompositingMode);
@@ -103,7 +103,7 @@ JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isColor
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setAlphaWriteEnable
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jboolean aEnable)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jboolean aEnable)
 {
     M3G_DO_LOCK
     m3gSetAlphaWriteEnable((M3GCompositingMode)aHCompositingMode, (M3Gbool)aEnable);
@@ -111,7 +111,7 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setAlphaWri
 }
 
 JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isDepthWriteEnabled
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jboolean enabled = (jboolean)m3gIsDepthWriteEnabled((M3GCompositingMode)aHCompositingMode);
@@ -120,7 +120,7 @@ JNIEXPORT jboolean JNICALL Java_javax_microedition_m3g_CompositingMode__1isDepth
 }
 
 JNIEXPORT jfloat JNICALL Java_javax_microedition_m3g_CompositingMode__1getDepthOffsetFactor
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jfloat factor = (jfloat)m3gGetDepthOffsetFactor((M3GCompositingMode)aHCompositingMode);
@@ -129,7 +129,7 @@ JNIEXPORT jfloat JNICALL Java_javax_microedition_m3g_CompositingMode__1getDepthO
 }
 
 JNIEXPORT jfloat JNICALL Java_javax_microedition_m3g_CompositingMode__1getDepthOffsetUnits
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jfloat offset = (jfloat)m3gGetDepthOffsetUnits((M3GCompositingMode)aHCompositingMode);
@@ -138,7 +138,7 @@ JNIEXPORT jfloat JNICALL Java_javax_microedition_m3g_CompositingMode__1getDepthO
 }
 
 JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setBlending
-(JNIEnv* aEnv, jclass, jint aHCompositingMode, jint aMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode, jint aMode)
 {
     M3G_DO_LOCK
     m3gSetBlending((M3GCompositingMode)aHCompositingMode, (int)aMode);
@@ -146,7 +146,7 @@ JNIEXPORT void JNICALL Java_javax_microedition_m3g_CompositingMode__1setBlending
 }
 
 JNIEXPORT jfloat JNICALL Java_javax_microedition_m3g_CompositingMode__1getAlphaThreshold
-(JNIEnv* aEnv, jclass, jint aHCompositingMode)
+(JNIEnv* aEnv, jclass, jlong aHCompositingMode)
 {
     M3G_DO_LOCK
     jfloat treshold = (jfloat)m3gGetAlphaThreshold((M3GCompositingMode)aHCompositingMode);
